app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==7.0.2
redis==5.2.1
